+++ ResidentVulnerabilitiesDataPipeline.Tests/Handler.cs
@@ -9,9 +9,9 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using Moq;
 using Npgsql;
 using NUnit.Framework;
-using S3ToPostgresDataPipeline.Database;
+using ResidentVulnerabilitiesDataPipeline.Database;
 
-namespace S3ToPostgresDataPipeline.Tests
+namespace ResidentVulnerabilitiesDataPipeline.Tests
 {
     [TestFixture]
     public class HandlerTest : DatabaseTests
