+++ ResidentVulnerabilitiesDataPipeline.Tests/ResidentVulnerabilitiesDataPipeline.Tests.csproj
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\S3ToPostgresDataPipeline\S3ToPostgresDataPipeline.csproj" />
+    <ProjectReference Include="..\ResidentVulnerabilitiesDataPipeline\ResidentVulnerabilitiesDataPipeline.csproj" />
   </ItemGroup>
 
 </Project>
