+++ ResidentVulnerabilitiesDataPipeline/build.cmd
@@ -1,0 +1,2 @@
+dotnet restore
+dotnet lambda package --configuration release --framework netcoreapp3.1 --output-package bin/release/netcoreapp3.1/ResidentVulnerabilitiesDataPipeline.zip
