--- S3ToPostgresDataPipeline/build.cmd
@@ -1,2 +1,0 @@
-dotnet restore
-dotnet lambda package --configuration release --framework netcoreapp3.1 --output-package bin/release/netcoreapp3.1/S3ToPostgresDataPipeline.zip
