+++ database.env
@@ -1,4 +1,4 @@
 DB_PORT=5432
-POSTGRES_DB=s3-to-postgres-data-pipeline-test-db
+POSTGRES_DB=resident-vulnerabilities-data-pipeline-test-db
 POSTGRES_USERNAME=postgres
 POSTGRES_PASSWORD=password
