+++ docker-compose.yml
@@ -1,10 +1,10 @@
 version: "3.2"
 
 services:
-  s3-to-postgres-data-pipeline:
-    image: s3-to-postgres-data-pipeline
+  resident-vulnerabilities-data-pipeline:
+    image: resident-vulnerabilities-data-pipeline
     build:
-      context: S3ToPostgresDataPipeline/
+      context: ResidentVulnerabilitiesDataPipeline/
       dockerfile: ./Dockerfile
     ports:
       - 3000:3000
@@ -19,11 +19,11 @@ services:
       - database.env
     image: postgres:12
 
-  s3-to-postgres-data-pipeline-test:
-    image: s3-to-postgres-data-pipeline-test
+  resident-vulnerabilities-data-pipeline-test:
+    image: resident-vulnerabilities-data-pipeline-test
     build:
       context: .
-      dockerfile: S3ToPostgresDataPipeline.Tests/Dockerfile
+      dockerfile: ResidentVulnerabilitiesDataPipeline.Tests/Dockerfile
     env_file:
       - database.env
     environment:
@@ -35,6 +35,6 @@ services:
     ports:
       - 5432:5432
     environment:
-      - POSTGRES_DB=s3-to-postgres-data-pipeline-test-db
+      - POSTGRES_DB=resident-vulnerabilities-data-pipeline-test-db
       - POSTGRES_USERNAME=postgres
       - POSTGRES_PASSWORD=password
